refactor(boilerplate): reworked kebab-case to camelCase
