+++ boilerplate/app/models/Episode.ts
@@ -1,6 +1,6 @@
 import { Instance, SnapshotIn, SnapshotOut, types } from "mobx-state-tree"
-import { withSetPropAction } from "./helpers/with-set-prop-action"
-import { formatDate } from "../utils/format-date"
+import { withSetPropAction } from "./helpers/withSetPropAction"
+import { formatDate } from "../utils/formatDate"
 import { translate } from "../i18n"
 
 interface Enclosure {
