+++ boilerplate/app/models/helpers/setupRootStore.ts
@@ -2,7 +2,7 @@
  * This file is where we do "rehydration" of your RootStore from AsyncStorage.
  * This lets you persist your state between app launches.
  *
- * Navigation state persistence is handled in navigation-utilities.tsx.
+ * Navigation state persistence is handled in navigationUtilities.tsx.
  *
  * Note that Fast Refresh doesn't play well with this file, so if you edit this,
  * do a full refresh of your app instead.
