+++ boilerplate/app/models/index.ts
@@ -1,4 +1,4 @@
 export * from "./RootStore"
-export * from "./helpers/get-root-store"
-export * from "./helpers/use-stores"
-export * from "./helpers/setup-root-store"
+export * from "./helpers/getRootStore"
+export * from "./helpers/useStores"
+export * from "./helpers/setupRootStore"
