+++ boilerplate/app/screens/ErrorScreen/ErrorBoundary.tsx
@@ -32,8 +32,8 @@ export class ErrorBoundary extends Component<Props, State> {
     })
 
     // You can also log error messages to an error reporting service here
-    // This is a great place to put BugSnag, Sentry, Honeybadger, etc:
-    // reportErrorToCrashReportingService(error)
+    // This is a great place to put BugSnag, Sentry, crashlytics, etc:
+    // reportCrash(error)
   }
 
   // Reset the error back to null
