+++ test/vanilla/ignite-generate.test.ts
@@ -76,9 +76,9 @@ describe("ignite-cli generate", () => {
       `)
       expect(read(`${TEMP_DIR}/app/models/index.ts`)).toMatchInlineSnapshot(`
         "export * from \\"./RootStore\\"
-        export * from \\"./helpers/get-root-store\\"
-        export * from \\"./helpers/use-stores\\"
-        export * from \\"./helpers/setup-root-store\\"
+        export * from \\"./helpers/getRootStore\\"
+        export * from \\"./helpers/useStores\\"
+        export * from \\"./helpers/setupRootStore\\"
         export * from \\"./Pizza\\"
         "
       `)
@@ -128,9 +128,9 @@ describe("ignite-cli generate", () => {
       `)
       expect(read(`${TEMP_DIR}/app/models/index.ts`)).toMatchInlineSnapshot(`
         "export * from \\"./RootStore\\"
-        export * from \\"./helpers/get-root-store\\"
-        export * from \\"./helpers/use-stores\\"
-        export * from \\"./helpers/setup-root-store\\"
+        export * from \\"./helpers/getRootStore\\"
+        export * from \\"./helpers/useStores\\"
+        export * from \\"./helpers/setupRootStore\\"
         export * from \\"./PizzaStore\\"
         "
       `)
